+++ tests/databricks_tests/run_tests_notebook.py
@@ -21,7 +21,7 @@ import pytest
 # wd = os.getcwd()
 wd = pathlib.Path().resolve()
 print(f"Current working directory: {wd}")
-unit_test_folder = (wd / "../unit_tests").as_posix()
+unit_test_folder = (wd / "../unit_tests").resolve().as_posix()
 # COMMAND ----------
